requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
psycopg2-binary==2.9.9
pydantic==2.9.2
python-multipart==0.0.12